fix su recipe card and controller recipe

+++ frontend/src/app/components/form-ingredient/form-ingredient.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import {NgxSpinnerService} from 'ngx-spinner';
 import {MessageService} from 'primeng/api';
@@ -16,11 +16,14 @@ import {Table} from "primeng/table";
 export class FormIngredientComponent implements OnInit {
   @Input() id: string | null = null;
   @Input() showBreadcrumbs: boolean = true;
+  @Input() fromRecipe: boolean = false;
+  @Input() editMode: boolean = false;
+  @Input() ingredients: Ingredient[] | undefined;
+  @Output() addedIngredient = new EventEmitter<Ingredient>();
 
-  public editMode: boolean = false;
   public loading: boolean = true;
 
-  public model: Ingredient = new Ingredient();
+  public model: Ingredient | null = null;
 
   loadData: boolean = true;
 
@@ -68,7 +71,11 @@ export class FormIngredientComponent implements OnInit {
 
     this.initVal();
 
-    this.loadIngredients();
+    if (!this.fromRecipe) {
+      this.loadIngredients();
+    } else {
+      this._ingredients = this.ingredients || [];
+    }
 
     if (this.editMode) {
       this.getIngredient(this.route.snapshot.params['id'] || this.id);
@@ -79,7 +86,6 @@ export class FormIngredientComponent implements OnInit {
 
   initVal() {
     this.model = new Ingredient();
-    this.model.type = "";
 
     this.submitted = false;
     this.loading = false;
@@ -124,46 +130,50 @@ export class FormIngredientComponent implements OnInit {
   }
 
   onSubmit() {
-    this.submitted = true;
 
-    debugger;
+    this.submitted = true;
     this.loading = true;
 
-    if (this.editMode) {
-      this.ingredientService.update(this.model._id, this.model).subscribe({
-        next: (res) => {
-          this.messageService.add({
-            severity: 'success',
-            summary: 'Service Message',
-            detail: 'Via MessageService',
-          });
-        },
-        error: (e) => console.error(e),
-      });
-    } else {
+    if (this.model) {
+      if (this.editMode) {
+        this.ingredientService.update(this.model._id, this.model).subscribe({
+          next: (res) => {
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Service Message',
+              detail: 'Via MessageService',
+            });
+          },
+          error: (e) => console.error(e),
+        });
+      } else if (!this.fromRecipe) {
+        let data = {
+          name: this.model.name,
+          type: this.model.type,
+          quantity: this.model.quantity,
+        }
 
-      let data = {
-        name: this.model.name,
-        type: this.model.type,
-        quantity: this.model.quantity,
+        this.ingredientService.create(data, IngredientRef.TO_USER).subscribe({
+          next: (res) => {
+            console.log(res);
+            this._ingredients.push(res);
+
+            this.spinner.show();
+            this.initVal();
+            this.displayModal = false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Service Message',
+              detail: 'Via MessageService',
+            });
+          },
+          error: (e) => console.error(e),
+        });
+      } else {
+        this._ingredients.push(this.model);
+        this.addedIngredient.emit(this.model);
+        this.displayModal = false;
       }
-
-      this.ingredientService.create(data, IngredientRef.TO_USER).subscribe({
-        next: (res) => {
-          console.log(res);
-          this._ingredients.push(res);
-
-          this.spinner.show();
-          this.initVal();
-
-          this.messageService.add({
-            severity: 'success',
-            summary: 'Service Message',
-            detail: 'Via MessageService',
-          });
-        },
-        error: (e) => console.error(e),
-      });
     }
   }
 
@@ -172,7 +182,7 @@ export class FormIngredientComponent implements OnInit {
   }
 
   deleteIngredient(): void {
-    if (this.editMode) {
+    if (this.editMode && this.model) {
       this.ingredientService.delete(this.model._id).subscribe({
         next: (res) => {
           this.router.navigate(['/ingredients']);
@@ -182,8 +192,10 @@ export class FormIngredientComponent implements OnInit {
     }
   }
 
-  public setNameSelected(event: Ingredient) {
-    this.model.name = event.name;
+  public setNameSelected(event: Ingredient | null) {
+    if (this.model){
+      this.model.name =  event ? event.name : null;
+    }
   }
 
   public onChangeType() {
@@ -232,5 +244,17 @@ export class FormIngredientComponent implements OnInit {
   showModalDialog() {
     this.model = new Ingredient();
     this.displayModal = true;
+  }
+
+  isDisabled() {
+    if (this.model &&
+      this.model.quantity && this.model.quantity > 0 &&
+      this.model.name && this.model.name !== "" &&
+      this.model.type && this.model.type !== ""
+    ) {
+      return false
+    }
+
+    return true;
   }
 }
